Refactor: moved get_color to it's own file
