+++ UnitTests/CMakeLists.txt
@@ -29,6 +29,7 @@ set(Headers
 set(Sources
 	../sources/validation/arg_validator.c
 	../sources/validation/scene_interpretation.c
+	../sources/validation/get_color.c
 )
 
 add_library(${This} ${Sources} ${Libft} ${Headers})
